Refactor ecosystem backend
