+++ apps/charterafrica/src/pages/api/v1/resources/ecosystem/[action].page.js
@@ -1,4 +1,4 @@
-import { schema } from "@/charterafrica/lib/ecosystem/airtable";
+import airtable from "@/charterafrica/lib/ecosystem/airtable";
 import {
   updateEcosystemList,
   updateEcosystemContent,
@@ -8,18 +8,32 @@ const isApiKeyValid = (key) => {
   return key && key === process.env.RESOURCES_SECRET_TOKEN;
 };
 
+async function schema(req, res) {
+  const { baseId } = req.query;
+  if (!baseId) {
+    return res.status(400).json({ message: "MISSING BASE ID" });
+  }
+  return airtable.schema(baseId);
+}
+
+async function bases() {
+  return airtable.bases();
+}
+
 const actionMap = {
   "update-list": updateEcosystemList,
   "update-content": updateEcosystemContent,
   schema,
+  bases,
 };
 
+const apiKeyExcludedActions = ["schema", "bases"];
 export default async function handler(req, res) {
   const {
     query: { action },
   } = req;
   const key = req.headers["x-api-key"];
-  if (!isApiKeyValid(key) && action !== "schema") {
+  if (!isApiKeyValid(key) && !apiKeyExcludedActions.includes(action)) {
     return res.status(403).json({ message: "INVALID_API_KEY" });
   }
   const actionFunc = actionMap[action];
